requirements.txt
google-genai
chromadb
